replaced delay() with millis() to fix slow down
I was experiencing about a quarter second slow down per minute, which is something I was afraid of with a clock, but I wanted to see how affected the program would be by just the delay function, and I decided a lot since it was easily noticeable, and the clock speed is not that high so it is orders of magnitude higher than the noticeability of the same thing on a modern processor. So the long and the short of it is a millis() if loop now replaces delay(), it should also have rollover protection but that is yet to be seen.